feat: componentization of details screen
