+++ lib/details/widgets/details_app_bar.dart
@@ -1,17 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:projeto_criptos/details/controller/x_axis_provider.dart';
 
-class DetailsAppBar extends StatelessWidget implements PreferredSizeWidget {
+class DetailsAppBar extends HookConsumerWidget implements PreferredSizeWidget {
   const DetailsAppBar({
     Key? key,
   }) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return AppBar(
       elevation: 1,
       leading: IconButton(
         onPressed: () {
           Navigator.of(context).pop();
+          ref.read(xAxisProvider.state).state = 5;
         },
         icon: const Icon(Icons.arrow_back),
         color: Colors.black,
